Added some ui changes
