+++ app/event/[id]/page.tsx
@@ -28,6 +28,7 @@ import SubHeader from "@/components/SubHeader/SubHeader";
 import { FaImage, FaInfoCircle } from "react-icons/fa";
 import { IoIosSend } from "react-icons/io";
 import { SingleImageDropzone } from "@/components/singledropZone/SingleImageDropZone";
+import WhatsOnYourMind from "@/components/WhatsOnYourMInd/WhatsOnYourMind";
 
 
 interface EventUpdateType {
@@ -514,82 +515,5 @@ const PrizeCard = ({ position, content }: { position: string, content: string })
     <HiMiniTrophy className="h-20 w-20"></HiMiniTrophy>
   </div>
 )
-
-const WhatsOnYourMind = () => {
-  const [file, setFile] = useState<File>();
-  const { edgestore } = useEdgeStore();
-  const [caption, setCaption] = useState<string>("");
-  const [progress, setProgress] = useState<number>(0);
-  const { user, isLoaded } = useUser();
-  const [mongoId, setMongoId] = useState<string | null>(null);
-  const [userName, setUserName] = useState<string | null>("")
-
-  useEffect(() => {
-    if (user) {
-      setMongoId(user.publicMetadata.mongoId as string)
-      setUserName(user.fullName)
-    }
-  }, [isLoaded, user])
-
-  const handlePost = () => {
-    console.log("clicked");
-
-    const post = async () => {
-      if (file) {
-        const response = await edgestore.mypublicImages.upload({
-          file,
-          onProgressChange: (progress) => {
-            setProgress(progress);
-          },
-        });
-
-        if (response.url) {
-          const res = await fetch(`/api/userpost`, {
-            method: "POST",
-            body: JSON.stringify({ image: response.url, caption, createdBy: user?.publicMetadata.mongoId, name: userName }),
-          })
-          if (res.ok) {
-            toast.success("Posted successfully")
-          }
-        }
-      }
-    }
-
-    console.log("before");
-
-    post();
-    console.log("posted");
-  }
-
-  return (<div className="bg-slate-900 rounded-xl w-full p-4 max-w-[600px] mx-auto mb-10 h-fit flex flex-col gap-5">
-    <textarea value={caption} onChange={(e: any) => setCaption(e.target.value)} placeholder="What's on your mind ?" className="placeholder:opacity-80 text-cyan-300 rounded-[2rem] py-3 px-4 w-full bg-transparent border-2 border-cyan-600"></textarea>
-    <div className="flex justify-between p-2">
-      <div className="flex">
-        <Dialog>
-          <DialogTrigger asChild>
-            <FaImage className="cursor-pointer w-7 h-7 ml-2 fill-cyan-500 hover:fill-cyan-300"></FaImage>
-          </DialogTrigger>
-          <DialogContent className="bg-slate-950 opacity-75">
-            <DialogTitle>Select Image</DialogTitle>
-            <div className="flex justify-center">
-              <SingleImageDropzone width={200}
-                height={200}
-                value={file}
-                onChange={(file) => {
-                  setFile(file);
-                }}></SingleImageDropzone>
-            </div>
-          </DialogContent>
-        </Dialog>
-
-      </div>
-      <button onClick={handlePost} className="hover:border-cyan-400 hover:text-cyan-200 p-2 w-fit flex gap-3 items-center self-end px-5 border-2 border-cyan-600 rounded-lg">
-        <IoIosSend className="fill-cyan-600 mt-[1px]"></IoIosSend>
-        <span className="text-cyan-400 text-lg">Post</span>
-      </button>
-    </div>
-
-  </div >)
-}
 
 export default EventPage;
